+++ Source/Core/VideoBackends/D3D12/main.cpp
@@ -140,10 +140,6 @@ void VideoBackend::InitBackendInfo()
   }
   factory->Release();
 
-  // Clear ppshaders string vector
-  g_Config.backend_info.PPShaders.clear();
-  g_Config.backend_info.AnaglyphShaders.clear();
-
   D3D::UnloadD3D();
   D3D::UnloadDXGI();
 }
