+++ Source/Core/VideoCommon/VideoConfig.h
@@ -170,8 +170,6 @@ struct VideoConfig final
 
     std::vector<std::string> Adapters;  // for D3D
     std::vector<int> AAModes;
-    std::vector<std::string> PPShaders;        // post-processing shaders
-    std::vector<std::string> AnaglyphShaders;  // anaglyph shaders
 
     // TODO: merge AdapterName and Adapters array
     std::string AdapterName;  // for OpenGL
